The user can now click a parent to toggle both its children and the links to them.
